Day 5 part 1 complete
What a doozy, took me a while and I needed to brush off
my rspec TDD chops in order to make sure i actually implemented
most of this correctly. What actually caught me the longest was
the immediate argument bit. Unfortunately I've chosen to pass in
the ENTIRE program into the instruction classes, which makes a bit
of sense, since they need to read arbitrary addresses, however it
meant that a read for the next parameter could have been very easily
incorrect, reading from the beginning of the array rather than whereverr
our evaluation index is. Might be nice to separate out these accesses,
so that when an instruction is read those next memory locations are
automatically read in. They shouldn't change within the course of
an instruction's evaluation, and if they do it's not until another
instruction begins to read those addresses that it matters.