Day 7 part 1
Extract intcode computer finally because it's clearly going to be used
more in the course of these exercises. Not that I'm complaining, it has
been a nice and refreshing experience to see the workings of a computer
system clarify themselves over the course of a few nights and with the
help of some clever friends that wrote these materials.